requirements.txt
streamlit==1.39.0
pandas==2.2.3
scikit-learn==1.5.1
plotly==5.24.1
xgboost==2.1.2
joblib==1.4.2
openai==1.54.4
python-dotenv==1.0.1
